Update servo driver and main application logic.

+++ drivers/servo.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 void servo_init(void);
-void servo_set_angle(uint8_t angle);   // 0–180°
+void servo_set_angle(uint16_t angle);   // 0–1080°
 
 #ifdef __cplusplus
 }
